Support execution of CREATE MATERIALIZED VIEW
Add createMaterializedView API to MetadataManager and add CreateMaterializedViewTask
that calls the API to perform MV creation.

HiveMetadata implementation of the API is to add essential MV table parameters
and create MV as a standard table. ConnectorMaterializedViewDefinition is the json
structure that contains all essential metadata we need to keep in Metastore. It will
serialize into a table parameter.

By this diff, querying MV works the same as querying a standard table. We only allow
query/select on MV and disallow all other statements, ex. create, insert, delete,
which is enforced in StatementAnalyzer.

We don't support MVs defined across different catalogs, which is enforced in
CreateMaterializedViewTask