Add materialized view freshness check to query
Check materialized view freshness during query.
Throw STALE_MATERIALIZED_VIEW exception if MV is not fresh.
MV is fresh if it is synced with the base tables, and all base table
partitions are propagated to MV, since it is created.

HiveMetastore implementation of freshness check is to compare
partition names of every base table with MV partition names, and
check if there is any MV partition missing.

This implementation assumes that, in the MV write/loading path,
a MV partition gets dropped if any corresponding base table partition
lands or is reloaded.